rag-collection-basic/light-rag-app/pyproject.toml
[project]
name = "light-rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "lightrag-hku>=1.4.9.1",
    "ollama>=0.6.0",
    "openai>=2.2.0",
]
